Fix bugprone-narrowing-conversions warnings in core/vnl
Add explicit static_cast<> conversions throughout core/vnl to resolve
all clang-tidy bugprone-narrowing-conversions warnings. Covers 35 files
including source, headers, and .hxx template implementations.

Fixes test failure: vul_vul_test_driver

Key patterns fixed:
- size_t/.size() → int: static_cast<int>()
- unsigned int → int at function call sites
- long → int in getter return values
- size_t → double in arithmetic expressions
- long double → double for complex number operations
- unsigned short arithmetic → short for vnl_bignum::resize() calls
- long (rational int_type) → double: replaced 0.5*den comparisons
  with integer arithmetic (2*(num%den) >= den)
- vnl_matrix.hxx vnl_inplace_transpose: added local int aliases
  im_/in_ for signed arithmetic on unsigned m/n parameters
- Sparse matrix inner loop vars changed from unsigned int to int
  (sparse_vector holds pair<int,int>)

+++ core/vnl/algo/vnl_amoeba.cxx
@@ -162,7 +162,7 @@ operator==(const vnl_amoeba_SimplexCorner & a, const vnl_amoeba_SimplexCorner & 
 void
 vnl_amoebaFit::set_up_simplex_relative(std::vector<vnl_amoeba_SimplexCorner> & simplex, const vnl_vector<double> & x)
 {
-  const int n = x.size();
+  const int n = static_cast<int>(x.size());
 
   simplex[0].v = x;
   simplex[0].fv = f(x);
@@ -192,7 +192,7 @@ vnl_amoebaFit::set_up_simplex_absolute(std::vector<vnl_amoeba_SimplexCorner> & s
                                        const vnl_vector<double> & x,
                                        const vnl_vector<double> & dx)
 {
-  const int n = x.size();
+  const int n = static_cast<int>(x.size());
 
   simplex[0].v = x;
   simplex[0].fv = f(x);
@@ -239,7 +239,7 @@ void
 vnl_amoebaFit::amoeba(vnl_vector<double> & x)
 {
   // Set up a simplex near the initial guess.
-  const int n = x.size();
+  const int n = static_cast<int>(x.size());
   std::vector<vnl_amoeba_SimplexCorner> simplex(n + 1, vnl_amoeba_SimplexCorner(n));
 
   set_up_simplex_relative(simplex, x);
@@ -250,7 +250,7 @@ void
 vnl_amoebaFit::amoeba(vnl_vector<double> & x, const vnl_vector<double> & dx)
 {
   // Set up a simplex near the initial guess.
-  const int n = x.size();
+  const int n = static_cast<int>(x.size());
   std::vector<vnl_amoeba_SimplexCorner> simplex(n + 1, vnl_amoeba_SimplexCorner(n));
 
   set_up_simplex_absolute(simplex, x, dx);
@@ -261,7 +261,7 @@ vnl_amoebaFit::amoeba(vnl_vector<double> & x, const vnl_vector<double> & dx)
 void
 vnl_amoebaFit::amoeba(vnl_vector<double> & x, std::vector<vnl_amoeba_SimplexCorner> & simplex)
 {
-  const int n = x.size();
+  const int n = static_cast<int>(x.size());
   sort_simplex(simplex);
 
   if (verbose > 1)
@@ -428,7 +428,7 @@ class vnl_amoeba_LSCF : public vnl_cost_function
 
 public:
   vnl_amoeba_LSCF(vnl_least_squares_function & ls)
-    : vnl_cost_function(ls.get_number_of_unknowns())
+    : vnl_cost_function(static_cast<int>(ls.get_number_of_unknowns()))
     , ls_(&ls)
     , fx(ls.get_number_of_residuals())
   {}
