+++ core/vnl/algo/vnl_gaussian_kernel_1d.cxx
@@ -26,7 +26,7 @@ vnl_gaussian_kernel_1d::vnl_gaussian_kernel_1d(double sigma, double cutoff)
   : vec_((int)std::ceil(compute_width(sigma, cutoff)))
   , inscale_(0.5 / (sigma * sigma))
 {
-  const int wid = vec_.size();
+  const int wid = static_cast<int>(vec_.size());
 
   double area = 0;
   for (int i = 0; i < wid; ++i)
